+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 ## Changelog
 
+### v1.1.1 (2024-11-11)
+
+#### Improvements and Ease of Use
+- adds a `class_index` parameter to `TabularExplainer` and `Explainer` to specify the class index to be explained for classification models [#271](https://github.com/mmschlk/shapiq/issues/271) (renames `class_label` parameter in TreeExplainer to `class_index`)
+- adds support for `PyTorch` models to `Explainer` [#272](https://github.com/mmschlk/shapiq/issues/272)
+
+#### Bug Fixes
+- fixes a bug that `RandomForestClassifier` models were not working with the `TreeExplainer` [#273](https://github.com/mmschlk/shapiq/issues/273)
+
 ### v1.1.0 (2024-11-07)
 
 #### New Features and Improvements